implemented add email types
