+++ src/components/Collections/CollectionShare/emailShare/sharees/row.tsx
