+++ src/constants/collection/collectionShare.tsx
@@ -1,0 +1,4 @@
+export enum EmailShareType {
+    viewer = 'viewer',
+    collaborator = 'collaborator',
+}
